почистил математические преблуды
